Alice's updates
Small changes to wording, plus suggestion that they submit links to their public profile on FCC to the class Slack channel before the first Sunday session. This can be recorded, like the first homework. Just came up with this now, so open to critiques! 
Also suggest making a Loom video on how to make the FCC profile public, but didn't have time to do it right now.